+++ packages/plugin-packages/mashroom-security-provider-ldap/src/fix_tls_options.ts
@@ -1,30 +1,30 @@
-// @flow
 
 import fs from 'fs';
 import path from 'path';
 
+import type {TlsOptions as NodeTlsOptions} from 'tls';
 import type {MashroomLoggerFactory} from '@mashroom/mashroom/type-definitions';
-import type {NodeTlsOptions, TlsOptions} from '../type-definitions';
+import type {TlsOptions} from '../type-definitions';
 
-const CERT_PROPERTIES = ['cert', 'ca', 'key', 'crl', 'pfx'];
+const CERT_PROPERTIES: Array<keyof TlsOptions> = ['cert', 'ca', 'key', 'crl', 'pfx'];
 
-export default (optionalTlsOptions: ?TlsOptions, serverRootFolder: string, loggerFactory: MashroomLoggerFactory): ?NodeTlsOptions => {
+export default (optionalTlsOptions: TlsOptions | undefined | null, serverRootFolder: string, loggerFactory: MashroomLoggerFactory): NodeTlsOptions | undefined | null => {
     if (!optionalTlsOptions) {
         return optionalTlsOptions;
     }
 
     const tlsOptions: TlsOptions = optionalTlsOptions;
-    const fixedTlsOptions: NodeTlsOptions = {...(tlsOptions: any)};
+    const fixedTlsOptions: any = {
+        ...tlsOptions
+    };
     const logger = loggerFactory('mashroom.security.provider.ldap');
 
     CERT_PROPERTIES.forEach((certPropName) => {
         if (tlsOptions.hasOwnProperty(certPropName)) {
-            let certs = tlsOptions[certPropName];
+            const certs = tlsOptions[certPropName] as string | Array<string>;
             if (certs) {
-                if (!Array.isArray(certs)) {
-                    certs = [certs];
-                }
-                certs = certs.map((certPath) => {
+                const certArray = Array.isArray(certs) ? certs : [certs];
+                let fixedCertArray: Array<string | Buffer | null> = certArray.map((certPath) => {
                     if (!path.isAbsolute(certPath)) {
                         certPath = path.resolve(serverRootFolder, certPath);
                     }
@@ -36,14 +36,13 @@ export default (optionalTlsOptions: ?TlsOptions, serverRootFolder: string, logge
                     }
                 });
 
-                certs = certs.filter((cert) => cert !== null);
-
-                fixedTlsOptions[certPropName] = (certs: any);
+                fixedCertArray = fixedCertArray.filter((cert) => cert !== null);
+                fixedTlsOptions[certPropName] = fixedCertArray;
             }
         }
     });
 
-    logger.debug('LDAP TLS options: ', {tlsOptions: fixedTlsOptions});
+    logger.debug('LDAP TLS options: ', { tlsOptions: fixedTlsOptions });
     return fixedTlsOptions;
 }