+++ packages/plugin-packages/mashroom-security-provider-ldap/src/mashroom-bootstrap.ts
@@ -1,4 +1,3 @@
-// @flow
 
 import MashroomLdapSecurityProvider from './MashroomLdapSecurityProvider';
 import LdapClientImpl from './LdapClientImpl';
@@ -9,14 +8,19 @@ import type {MashroomSecurityProviderPluginBootstrapFunction} from '@mashroom/ma
 const bootstrap: MashroomSecurityProviderPluginBootstrapFunction = async (pluginName, pluginConfig, pluginContextHolder) => {
     const {
         loginPage, userSearchFilter, groupSearchFilter, groupToRoleMapping: groupToRoleMappingPath,
-        serverUrl, tlsOptions, baseDN, bindDN, bindCredentials, authenticationTimeoutSec,
+        serverUrl, ldapConnectTimeout, ldapTimeout,
+        tlsOptions, baseDN, bindDN, bindCredentials, authenticationTimeoutSec,
     } = pluginConfig;
 
     const pluginContext = pluginContextHolder.getPluginContext();
     const { loggerFactory, serverConfig: { serverRootFolder } } = pluginContext;
 
     const fixedTlsOptions = fixTlsOptions(tlsOptions, serverRootFolder, loggerFactory);
-    const ldapClient = new LdapClientImpl(serverUrl, baseDN, bindDN, bindCredentials, fixedTlsOptions, loggerFactory);
+    const ldapClient = new LdapClientImpl(serverUrl, ldapConnectTimeout, ldapTimeout, baseDN, bindDN, bindCredentials, fixedTlsOptions, loggerFactory);
+
+    pluginContext.services.core.pluginService.onUnloadOnce(pluginName, () => {
+       ldapClient.shutdown();
+    });
 
     return new MashroomLdapSecurityProvider(loginPage, userSearchFilter, groupSearchFilter, groupToRoleMappingPath, ldapClient, serverRootFolder, authenticationTimeoutSec, loggerFactory);
 };